feat(cli): add unified bootnode command
