+++ bin/base/src/commands/mod.rs
@@ -1,0 +1,6 @@
+//! Top-level command implementations for the unified Base binary.
+
+mod bootnode;
+mod command;
+pub(crate) use command::BaseCommand;
+mod rpc;
